requirements.txt
altgraph==0.17.2
fonttools==4.25.0
macholib==1.16
MouseInfo==0.1.3
munkres==1.1.4
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
PyGetWindow==0.0.9
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
rubicon-objc==0.4.2
